OSh/feature permission policies

+++ app/Http/Controllers/AttachmentController.php
@@ -61,6 +61,8 @@ class AttachmentController extends ApiController
      */
     public function storeTopicAttachment(Topic $topic, Request $request)
     {
+        $this->authorize('createTopicAttachment', $topic);
+
         $attachment_data = AttachmentService::uploadAttachmentToCloud($request);
         $attachment = Attachment::create($attachment_data);
         $attachment = $topic->attachments()->save($attachment);
@@ -76,6 +78,9 @@ class AttachmentController extends ApiController
     public function destroyTopicAttachment(Topic $topic, Attachment $attachment)
     {
         if ($this->isAttachmentBelongsToTopic($topic, $attachment)) {
+
+            $this->authorize('deleteTopicAttachment', $topic);
+
             AttachmentService::deleteAttachmentFromCloud($attachment->cloud_public_id);
             $attachment->delete();
             return $this->setStatusCode(204)->respond();
@@ -129,6 +134,8 @@ class AttachmentController extends ApiController
      */
     public function storeVoteAttachment(Vote $vote, Request $request)
     {
+        $this->authorize('createVoteAttachment', $vote);
+
         $attachment_data = AttachmentService::uploadAttachmentToCloud($request);
         $attachment = Attachment::create($attachment_data);
         $attachment = $vote->attachments()->save($attachment);
@@ -144,6 +151,9 @@ class AttachmentController extends ApiController
     public function destroyVoteAttachment(Vote $vote, Attachment $attachment)
     {
         if ($this->isAttachmentBelongsToVote($vote, $attachment)) {
+
+            $this->authorize('deleteVoteAttachment', $vote);
+
             AttachmentService::deleteAttachmentFromCloud($attachment->cloud_public_id);
             $attachment->delete();
             return $this->setStatusCode(204)->respond();
@@ -195,6 +205,8 @@ class AttachmentController extends ApiController
      */
     public function storeCommentAttachment(Comment $comment, Request $request)
     {
+        $this->authorize('createCommentAttachment', $comment);
+
         $attachment_data = AttachmentService::uploadAttachmentToCloud($request);
         $attachment = Attachment::create($attachment_data);
         $attachment = $comment->attachments()->save($attachment);
@@ -210,6 +222,9 @@ class AttachmentController extends ApiController
     public function destroyCommentAttachment(Comment $comment, Attachment $attachment)
     {
         if ($this->isAttachmentBelongsToComment($comment, $attachment)) {
+
+            $this->authorize('deleteCommentAttachment', $comment);
+
             AttachmentService::deleteAttachmentFromCloud($attachment->cloud_public_id);
             $attachment->delete();
             return $this->setStatusCode(204)->respond();
